Silenced Compiler Warnings
I am not a fan of doing this, but I made the decision to put parenthases in on these two lines of code. as! forces a non-nil conversion; if the conversion can't happen, the app crashes. I'm okay with the app failing fast in this way, since it hasn't so far and a crash would force improvement on the app.